+++ storage/types.go
@@ -47,7 +47,7 @@ type Backend interface {
 	PublishVolume(
 		ctx context.Context, volConfig *VolumeConfig, publishInfo *utils.VolumePublishInfo,
 	) error
-	UnpublishVolume(ctx context.Context, volConfig *VolumeConfig, publishInfo *utils.VolumePublishInfo) error
+	UnpublishVolume(ctx context.Context, volConfig *VolumeConfig, nodes []*utils.Node) error
 	GetVolumeExternal(ctx context.Context, volumeName string) (*VolumeExternal, error)
 	ImportVolume(ctx context.Context, volConfig *VolumeConfig) (*Volume, error)
 	ResizeVolume(ctx context.Context, volConfig *VolumeConfig, newSize string) error